refactor connection-status domain
